First commit, smoke-tested working solution
This is a new Openfire plugin that deploys the third-party 'XMPP-web' webclient to the embedded webserver that's part of Openfire.

The upstream project can be found at https://github.com/nioc/xmpp-web

This commit is based on release 0.10.0 of the upstream project.

The plugin has been briefly tested: it deploys, and a demo-user can use it to login and join chatrooms.