fixes #3: Openfire 5.0.0 compatiblity
Updates the plugin to be compatible with Openfire 5.0.0 (in particular: with the new version of Jetty that's part of that release).

This drops compatibility with versions of Openfire prior to 5.0.0.

Drop unnecessary manual initialization of JasperInitializer

To quote Joakim Erdfelt, a Jetty developer:

> For a `WebAppContext`, this manual initialization of JasperInitializer is not necessary. (it is necessary for a `ServletContextHandler`)
> The fact that jetty-ee8-apache-jsp-<ver>.jar is in your classpath should be enough for that webapp to automatically have JSP support enabled.
>
> If you find that you don't have a `jsp` servlet initialized, then that means you'll want to declare the `WebAppContext.setDefaultDescriptor()` to point to the `webdefault-ee8.xml` for your version. You can find that file in the `jetty-ee8-webapp-<ver>-config.jar` -
> See https://repo1.maven.org/maven2/org/eclipse/jetty/ee8/jetty-ee8-webapp/12.0.14/

Drop explicit InstanceManager configuration

Removing initialization of an attribute that I believe to be something that has been carried over from earlier versions, without this being neccesary anymore.